Refactor rekordbox package
Why?
rewrite_xml belongs in the rekordbox package rather than the sync
package. Other modules in the rekordbox package are cluttered.

What?
Move sync.helpers.rewrite_xml to rekordbox.helpers.rewrite_xml. Move all
non-core module functions into a helpers module.
