+++ src/djtools/rekordbox/tag_parsers.py
@@ -15,94 +15,15 @@ from abc import ABC, abstractmethod
 from collections import defaultdict
 import logging
 import re
-from typing import Any, Dict, List, Optional, Set, Tuple, Union
+from typing import Dict, List, Optional, Set, Tuple, Union
 
 import bs4
 from bs4 import BeautifulSoup
 
+from djtools.rekordbox.helpers import BooleanNode
+
 
 logger = logging.getLogger(__name__)
-
-
-class BooleanNode:
-    """Node that contains boolean logic for a subexpression."""
-
-    def __init__(self, parent: Optional[Any] = None):
-        """Constructor.
-
-        Args:
-            parent: BooleanNode of which this node is a subexpression.
-        """
-        self.parent = parent
-        self.operators = []
-        self.tags = []
-        self.tracks = []
-    
-    def __call__(
-        self, tracks: Dict[str, List[Tuple[str, List[str]]]]
-    ) -> Set[str]:
-        """Evaluates the boolean algebraic subexpression.
-
-        Args:
-            tracks: Map of tags to dicts of track_id: tags.
-
-        Raises:
-            RuntimeError: Boolean expressions must be formatted correctly.
-
-        Returns:
-            Reduced set of track IDs.
-        """
-        operators = len(self.operators)
-        operands = len(self.tags) + len(self.tracks)
-        if operators + 1 != operands:
-            raise RuntimeError(
-                f"Invalid boolean expression: track sets: {len(self.tracks)}, "
-                f"tags: {self.tags}, operators: "
-                f"{[x.__name__ for x in self.operators]}"
-            )
-        while self.tags or self.operators:
-            operator = self.operators.pop(0)
-            if len(self.tracks):
-                tracks_A = self.tracks.pop(0)
-            else:
-                tracks_A = self._get_tag_tracks(
-                    tag=self.tags.pop(0), tracks=tracks
-                )
-            if len(self.tracks):
-                tracks_B = self.tracks.pop(0)
-            else:
-                tracks_B = self._get_tag_tracks(
-                    tag=self.tags.pop(0), tracks=tracks
-                )
-            self.tracks.insert(0, operator(tracks_A, tracks_B))
-        
-        return next(iter(self.tracks), set())
-
-    def _get_tag_tracks(
-        self, tag: str, tracks: Dict[str, List[Tuple[str, List[str]]]]
-    ) -> Set[str]:
-        """Gets set of track IDs for the provided tag.
-
-        If the tag contains a wildcard, denoted with "*", then the union of
-        track IDs with a tag containing the provided tag as a substring is
-        returned.
-
-        Args:
-            tag: Tag for indexing tracks.
-            tracks: Map of tags to dicts of track_id: tags.
-
-        Returns:
-            Set of track IDs for the provided tag.
-        """
-        if "*" in tag:
-            exp = re.compile(r".*".join(tag.split("*")))
-            track_ids = set()
-            for k in tracks:
-                if re.search(exp, k):
-                    track_ids.update(set(tracks[k].keys()))
-            return track_ids
-        else:
-            return set(tracks.get(tag, {}).keys())
 
 
 class TagParser(ABC):
